Update expected test data for X.509 certs and CSRs
The change is because mbedtls now uses UTF8STRING instead of PRINTABLESTRING
for the DN entries.

In addition this starts setting a serial number on the generated (test)
certificate. If you don't explicitly set the serial number, Mbedtls actually
generates an invalid certificate (with a zero-length integer field, which is not
valid DER). This can't be parsed by other tools like `openssl x509`, which makes
debugging awkward.