Clean up features
Features are now split into two categories: cargo features and
platform features. Platform features are dependent on the
compilation target and automatically selected when needed. This
removes the (broken) mechanism of selecting backend implementations
based on cargo features. Enabled platform features are passed from
`mbedtls-sys` to dependent crates so they can change their
implementation as needed.

As part of this change, it's no longer possible to build the
`mbedtls` crate without threading support. Support may be added
again in the future if there's a need for it. The `mbedtls-sys`
crate can still be built without threading support.

The changes to `mbedtls-sys` are intended to not be a breaking
change (except for the removal of the libc feature, which had
not been explicitly advertised to users).

Also some cleanup related to Rust 2018 edition.