Namespace symbols for mbedtls_malloc to avoid conflicts.
Dependents of this crate may use multiple different versions of this crate. These
different versions all define mbedtls_calloc and mbedtls_free, which can result in
linking conflicts. Since these functions are defined in and only used by this
crate, we can namespace the symbols to avoid conflicts. We use Cargo's crate
disambiguator for this purpose.